Enforce limit for grep
