+++ front/lib/actions/mcp_actions.ts
@@ -9,6 +9,11 @@ import assert from "assert";
 import EventEmitter from "events";
 import type { JSONSchema7 } from "json-schema";
 
+import {
+  calculateContentSize,
+  getMaxSize,
+  isValidContentSize,
+} from "@app/lib/actions/action_output_limits";
 import type { MCPToolStakeLevelType } from "@app/lib/actions/constants";
 import {
   DEFAULT_CLIENT_SIDE_MCP_TOOL_STAKE_LEVEL,
@@ -96,22 +101,7 @@ function isEmptyInputSchema(schema: JSONSchema7): boolean {
 
 const MAX_TOOL_NAME_LENGTH = 64;
 
-// Size limits for MCP tool outputs
-export const MAX_TEXT_CONTENT_SIZE = 2 * 1024 * 1024; // 2MB.
-export const MAX_IMAGE_CONTENT_SIZE = 2 * 1024 * 1024; // 2MB.
-export const MAX_RESOURCE_CONTENT_SIZE = 10 * 1024 * 1024; // 10MB.
-
-export const MAXED_OUTPUT_FILE_SNIPPET_LENGTH = 64_000; // Approximately 16K tokens.
-
 export const TOOL_NAME_SEPARATOR = "__";
-
-export function computeTextByteSize(text: string): number {
-  return text.length * 2; // UTF-8 approximate
-}
-
-export function computeBase64ByteSize(base64: string): number {
-  return Math.ceil((base64.length * 3) / 4);
-}
 
 // Define the new type here for now, or move to a dedicated types file later.
 export interface ServerToolsAndInstructions {
@@ -382,14 +372,6 @@ export async function* tryCallMCPTool(
       };
     }
 
-    const isValidContentSize = (
-      content: CallToolResult["content"]
-    ): boolean => {
-      return !content.some(
-        (item) => calculateContentSize(item) > getMaxSize(item)
-      );
-    };
-
     const generateContentMetadata = (
       content: CallToolResult["content"]
     ): {
@@ -409,49 +391,6 @@ export async function* tryCallMCPTool(
         }
       }
       return result;
-    };
-
-    const getMaxSize = (item: CallToolResult["content"][number]) => {
-      switch (item.type) {
-        case "text":
-          return MAX_TEXT_CONTENT_SIZE;
-        case "image":
-          return MAX_IMAGE_CONTENT_SIZE;
-        case "resource":
-          return MAX_RESOURCE_CONTENT_SIZE;
-        default:
-          return 1 * 1024 * 1024; // 1MB default
-      }
-    };
-
-    const calculateContentSize = (
-      item: CallToolResult["content"][number]
-    ): number => {
-      switch (item.type) {
-        case "text":
-          return computeTextByteSize(item.text);
-        case "image":
-          return computeBase64ByteSize(item.data);
-        case "resource":
-          if (
-            "blob" in item.resource &&
-            item.resource.blob &&
-            typeof item.resource.blob === "string"
-          ) {
-            return computeBase64ByteSize(item.resource.blob);
-          }
-          if (
-            "text" in item.resource &&
-            typeof item.resource.text === "string"
-          ) {
-            return computeTextByteSize(item.resource.text);
-          }
-          return 0;
-        case "audio":
-          return item.data.length;
-        default:
-          return 0;
-      }
     };
 
     const serverType = (() => {
